+++ app/utils/Security.scala
@@ -1,4 +1,4 @@
-package util
+package utils
 
 import controllers.routes
 import play.api.mvc._
